+++ BOM-API v2/Schemas/PatchInputFormats.cs
@@ -7,19 +7,19 @@ namespace BillOfMaterialsAPI.Schemas
     {
         [Required][MaxLength(25)] public string item_id { get; set; }
         [Required][MaxLength(3)][RegularExpression("^(" + IngredientType.Material + "|" + IngredientType.InventoryItem + ")$", ErrorMessage = "Value must be either IngredientType.Material or IngredientType.InventoryItem")] public string ingredient_type { get; set; }
-        [Required] public int amount { get; set; }
+        [Required] public double amount { get; set; }
         [Required][MaxLength(15)] public string amount_measurement { get; set; }
     }
     public class PatchMaterials
     {
         [Required][MaxLength(50)] public string material_name { get; set; }
-        [Required] public int amount { get; set; }
+        [Required] public double amount { get; set; }
         [Required][MaxLength(15)] public string amount_measurement { get; set; }
     }
     public class PatchMaterialIngredients
     {
         [Required][MaxLength(25)] public string item_id { get; set; }
-        [Required] public int amount { get; set; }
+        [Required] public double amount { get; set; }
         [Required][MaxLength(15)] public string amount_measurement { get; set; }
     }
     public class PatchPastryMaterials
